# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Gate — AI editorial quality control API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type Verdict = typeof Verdict[keyof typeof Verdict];


export const Verdict = {
  APPROVED: 'APPROVED',
  APPROVED_WITH_NOTES: 'APPROVED_WITH_NOTES',
  REVISE: 'REVISE',
  BLOCKED: 'BLOCKED',
} as const;

export type GateMode = typeof GateMode[keyof typeof GateMode];


export const GateMode = {
  content: 'content',
  strategy: 'strategy',
  technical: 'technical',
  communication: 'communication',
} as const;

export interface CriterionResult {
  pass: boolean;
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
  assessment: string;
  evidence: string;
}

export interface GateRunBody {
  /**
     * The content to evaluate
     * @minLength 1
     * @maxLength 50000
     */
  draft: string;
  /**
     * What was requested — the gate checks whether the output answers it
     * @minLength 1
     * @maxLength 5000
     */
  brief: string;
  /**
     * Brand voice guide. If omitted, voice compliance is scored as neutral.
     * @maxLength 10000
     */
  voice_guide?: string;
  mode?: GateMode;
}

export type GateVerdictCriteriaResults = {[key: string]: CriterionResult};

export interface RateLimitInfo {
  used: number;
  limit: number;
  reset_at: string;
}

export type GateVerdictMeta = {
  mode?: GateMode;
  draft_hash?: string;
  has_voice_guide?: boolean;
  rate_limit?: RateLimitInfo;
};

export interface GateVerdict {
  /** Unique review ID for retrieval */
  id: string;
  verdict: Verdict;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  weighted_score?: number;
  criteria_results: GateVerdictCriteriaResults;
  strengths: string[];
  gaps: string[];
  revision_instructions?: string | null;
  missing_context?: string | null;
  meta?: GateVerdictMeta;
}

export interface CreateGateKeyBody {
  /**
     * Human-readable name for this key
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /**
     * Maximum gate evaluations per hour for this key
     * @minimum 1
     * @maximum 1000
     */
  rate_limit_per_hour?: number;
}

export interface CreateGateKeyResponse {
  id: string;
  /** Full API key — shown only once, store securely */
  key: string;
  /** Key prefix used for identification and revocation */
  prefix: string;
  name: string;
  rate_limit_per_hour: number;
  created_at: string;
  note?: string;
}

export type GateKeyRecordStatus = typeof GateKeyRecordStatus[keyof typeof GateKeyRecordStatus];


export const GateKeyRecordStatus = {
  active: 'active',
  revoked: 'revoked',
} as const;

export interface GateKeyRecord {
  id: string;
  prefix: string;
  name: string;
  rate_limit_per_hour: number;
  created_at: string;
  revoked_at?: string | null;
  status: GateKeyRecordStatus;
}

export type RevokeKeyResponseStatus = typeof RevokeKeyResponseStatus[keyof typeof RevokeKeyResponseStatus];


export const RevokeKeyResponseStatus = {
  revoked: 'revoked',
} as const;

export interface RevokeKeyResponse {
  prefix: string;
  status: RevokeKeyResponseStatus;
  revoked_at: string;
}

export type GateReviewRecordScorecard = {[key: string]: CriterionResult};

export type GateReviewRecordMeta = { [key: string]: unknown };

export interface GateReviewRecord {
  id: string;
  verdict: Verdict;
  confidence: number;
  weighted_score?: number | null;
  scorecard: GateReviewRecordScorecard;
  strengths: string[];
  gaps: string[];
  revision_instructions?: string | null;
  missing_context?: string | null;
  draft_hash: string;
  meta?: GateReviewRecordMeta;
  created_at: string;
}

export type GateReviewListPagination = {
  limit: number;
  offset: number;
  count: number;
};

export interface GateReviewList {
  data: GateReviewRecord[];
  pagination: GateReviewListPagination;
}

export type ErrorResponseDetailsItem = { [key: string]: unknown };

export interface ErrorResponse {
  error: string;
  message: string;
  details?: ErrorResponseDetailsItem[];
}

/**
 * Invalid request body or parameters
 */
export type ValidationErrorResponse = ErrorResponse;

/**
 * Missing or invalid API key
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Invalid admin secret
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Conflict with current state
 */
export type ConflictResponse = ErrorResponse;

export type RateLimitExceededResponse = ErrorResponse & {
  rate_limit?: RateLimitInfo;
};

/**
 * Internal server error
 */
export type InternalErrorResponse = ErrorResponse;

/**
 * LLM service error
 */
export type LLMErrorResponse = ErrorResponse;

export type ListGateReviewsParams = {
/**
 * Number of results to return
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * Number of results to skip
 * @minimum 0
 */
offset?: number;
/**
 * Filter by verdict
 */
verdict?: Verdict;
};