# lib/api-zod/src/generated/types/rateLimitInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Gate — AI editorial quality control API
 * OpenAPI spec version: 0.2.0
 */

export interface RateLimitInfo {
  used: number;
  limit: number;
  reset_at: Date;
}