# lib/api-zod/src/generated/types/verdict.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Gate — AI editorial quality control API
 * OpenAPI spec version: 0.2.0
 */

export type Verdict = typeof Verdict[keyof typeof Verdict];


export const Verdict = {
  APPROVED: 'APPROVED',
  APPROVED_WITH_NOTES: 'APPROVED_WITH_NOTES',
  REVISE: 'REVISE',
  BLOCKED: 'BLOCKED',
} as const;